+++ nab-kafka/src/main/java/ru/hh/nab/kafka/consumer/AckProvider.java
@@ -4,5 +4,5 @@ import org.apache.kafka.clients.consumer.Consumer;
 
 @FunctionalInterface
 interface AckProvider<T> {
-  Ack<T> createAck(KafkaConsumer<T> kafkaConsumer, Consumer<?, ?> nativeKafkaConsumer, RetryService<T> retryService);
+  Ack<T> createAck(KafkaConsumer<T> kafkaConsumer, Consumer<?, ?> nativeKafkaConsumer);
 }